cleanup pull request
