+++ src/Invoice/Calculator/ShortInvoiceCalculator.php
@@ -16,7 +16,7 @@ use App\Invoice\CalculatorInterface;
  * A calculator that sums up all timesheet records from the model and returns only one
  * entry for a compact invoice version.
  */
-class ShortInvoiceCalculator extends AbstractCalculator implements CalculatorInterface
+class ShortInvoiceCalculator extends AbstractMergedCalculator implements CalculatorInterface
 {
     /**
      * @return Timesheet[]
@@ -35,42 +35,6 @@ class ShortInvoiceCalculator extends AbstractCalculator implements CalculatorInt
         }
 
         return [$timesheet];
-    }
-
-    /**
-     * @param Timesheet $timesheet
-     * @param Timesheet $entry
-     */
-    protected function mergeTimesheets(Timesheet $timesheet, Timesheet $entry)
-    {
-        $timesheet->setUser($entry->getUser());
-        $timesheet->setFixedRate($entry->getFixedRate()); // FIXME invoice
-        $timesheet->setHourlyRate($entry->getHourlyRate()); // FIXME invoice
-        $timesheet->setRate($timesheet->getRate() + $entry->getRate());
-        $timesheet->setDuration($timesheet->getDuration() + $entry->getDuration());
-
-        if (null == $timesheet->getBegin() || $timesheet->getBegin()->getTimestamp() > $entry->getBegin()->getTimestamp()) {
-            $timesheet->setBegin($entry->getBegin());
-        }
-
-        if (null == $timesheet->getEnd() || $timesheet->getEnd()->getTimestamp() < $entry->getEnd()->getTimestamp()) {
-            $timesheet->setEnd($entry->getEnd());
-        }
-
-        if (null !== $this->model->getQuery()->getActivity()) {
-            $timesheet->setActivity($this->model->getQuery()->getActivity());
-            $timesheet->setDescription($this->model->getQuery()->getActivity()->getName());
-        } elseif (null !== $this->model->getQuery()->getProject()) {
-            $timesheet->setDescription($this->model->getQuery()->getProject()->getName());
-        }
-
-        if (null === $timesheet->getActivity()) {
-            $timesheet->setActivity($entry->getActivity());
-        }
-        if (empty($timesheet->getDescription())) {
-
-            $timesheet->setDescription($entry->getActivity()->getName());
-        }
     }
 
     /**
