+++ src/Invoice/Calculator/UserInvoiceCalculator.php
@@ -15,7 +15,7 @@ use App\Invoice\CalculatorInterface;
 /**
  * A calculator that sums up the timesheet records by user.
  */
-class UserInvoiceCalculator extends AbstractCalculator implements CalculatorInterface
+class UserInvoiceCalculator extends AbstractMergedCalculator implements CalculatorInterface
 {
     /**
      * @return Timesheet[]
@@ -39,41 +39,6 @@ class UserInvoiceCalculator extends AbstractCalculator implements CalculatorInte
         }
 
         return array_values($timesheets);
-    }
-
-    /**
-     * @param Timesheet $timesheet
-     * @param Timesheet $entry
-     */
-    protected function mergeTimesheets(Timesheet $timesheet, Timesheet $entry)
-    {
-        $timesheet->setUser($entry->getUser());
-        $timesheet->setFixedRate($entry->getFixedRate()); // FIXME invoice
-        $timesheet->setHourlyRate($entry->getHourlyRate()); // FIXME invoice
-        $timesheet->setRate($timesheet->getRate() + $entry->getRate());
-        $timesheet->setDuration($timesheet->getDuration() + $entry->getDuration());
-
-        if (null === $timesheet->getBegin() || $timesheet->getBegin()->getTimestamp() > $entry->getBegin()->getTimestamp()) {
-            $timesheet->setBegin($entry->getBegin());
-        }
-
-        if (null === $timesheet->getEnd() || $timesheet->getEnd()->getTimestamp() < $entry->getEnd()->getTimestamp()) {
-            $timesheet->setEnd($entry->getEnd());
-        }
-
-        if (null !== $this->model->getQuery()->getActivity()) {
-            $timesheet->setActivity($this->model->getQuery()->getActivity());
-            $timesheet->setDescription($this->model->getQuery()->getActivity()->getName());
-        } elseif (null !== $this->model->getQuery()->getProject()) {
-            $timesheet->setDescription($this->model->getQuery()->getProject()->getName());
-        }
-
-        if (null === $timesheet->getActivity()) {
-            $timesheet->setActivity($entry->getActivity());
-        }
-        if (empty($timesheet->getDescription())) {
-            $timesheet->setDescription($entry->getActivity()->getName());
-        }
     }
 
     /**
