+++ templates/invoice/index.html.twig
@@ -37,7 +37,6 @@
 {% endblock %}
 
 {% block javascripts %}
-    {# we do not call parent() as we use a custom built for the frontend assets and don't want the default <script> #}
     <script type="text/javascript">
         $(document).ready(function () {
             $( "#print-invoice-button" ).click(function() {
